requirements.txt
flask
pyyaml
RPi.GPIO